Simplify set up for local dev
Uses the `-free` version of datomic in the default :dev profile,
as well as starting in the `user` namespace, which seems to be
the one in the `dev/` folder, rather than `env/dev` (not sure what
the duplication is for).

Includes a few new functions in the `user` namespace to assist with
quick-start:

* (init-database) gets the ball rolling by creating the database and
    installing the schemas. Defaults to :free mode for simplicity,
    but if you have the license you can provide :dev
* (start-server) does what you would think, and (stop-server) undoes it

Also, if there's no :datomic-url environment variable set, (system :dev)
defaults to the :free variant that would be created by (init-database).
(init-database) is for quick-start anyway, so if you need a totally
different URI it's not too far out of your way to init the db by hand
and provide the environment variable.