Add (verify-new-user) `user` function, plus some more helpers
Since the local app doesn't seem to send emails, there's no easy
way to verify new users on a local build. This function solves
that by accepting either the username or the email and doing
everything automatically.

Includes a (with-db) function that is helpful for exploring the
`routes` functions, and fixes the (fig-start) function in the
`user` namespace, as well.