Add profile to automatically start server
Adds :start-server profile that will automatically boot up server when you run `lein with-profile +start-server repl`

Also switches Figwheel to open up the site on port 8890. Since it doesn't currently work with Pedestal interceptors, it cannot act as a fully-fledged server for all routes. To compensate, many routes are silently redirected to port 8890. This works, but is a bit flaky. Any hiccups in CORS, and the site will break. Since the web server runs off port 8890, we should default to that (Figwheel will continue to use port 3449 for its own WebSocket, which is fine)